refactor(apps): split the screen program off its compiler loader, so /edge carries no node builtin
`screen-tsc.ts` opened with `import { createRequire } from "node:module"`, and
`server/edge/typecheck.ts` reaches that file for `screenProgramWith`. Under
workerd without nodejs_compat the whole worker then refuses to START — `No such
module "node:module"`, at module-graph instantiation, before a request and
before any of the three primitives is reached.

Nothing calling it was never the question: esbuild keeps the external import
statement even after tree-shaking every consumer away, `sideEffects: false` does
not change that, and with node builtins not external the bundle will not build
at all. A lazy import would have made the outcome depend on bundler behaviour
and on whichever externals a consumer configured. A module boundary cannot be
defeated by either.

So the program builder, the compiler host, the options and the diagnostic
translation move to ./screen-program.ts, which imports no builtin.
`screen-tsc.ts` keeps exactly what needs a Node — finding the compiler through
`createRequire`, and reading the lib files off disk — and calls into it.
Reorganisation only: `screenProgram`, `loadCompiler`, `usable` and the
wire-screen `screenTscFindings` path keep their semantics and their sentences,
and the suites are unchanged.

Verified under real workerd with `compatibilityFlags: []`: all three primitives
run and the screen paints ["One","Two"].
